split main loop into two threads and modify raylib
The network polling is delayed if run on the same thread as
graphics which then blocks waiting for vsync. Put most of
the game on a second thread while leaving graphics on the
main thread to or not block as it needs. This requires modifying
raylib to allow calling SwapBuffers and PollInputEvents separately.
Because PollInputEvents should be called while the graphics
thread (main thread here) has exclusive access to the engine
while SwapBuffers must be called when the exclusive lock is
not held.

A side effect of this change is it is pretty straightforward
to have a dedicated server with no graphics.