new level collision code
each block has a profile as seen from above which can be sampled.
a profile is either a base profile or flipped somehow.

when probing from directions other than above the profile is rotated.

when a particle-like thing wants to move it probes in the 2 axes it
is traveling in to see how far it can get. If a surface is detected
it depenetrated if necessary and a collision is signaled. Otherwise
it is moved the entire way. A surface normal is returned when a
surface is detected.

the LooseItem reacts to collisions by reflecting the velocity using
the normal.