fix a bug hitting steep slopes from vertical
the depenetration happened, but it's not enough to get away
from the slope. Because Y velocity is still enough to go back
into the slope. Because the collision reaction didn't happen yet.

So in the collision reaction we move the item 1 step so it goes
away from the wall before the next collision call.